requirements.txt
numpy==1.12.1
opencv-python==3.2.0.7
requests==2.13.0
Pillow==4.1.1
git+https://github.com/python-telegram-bot/python-telegram-bot.git